+++ src/Crc16.cpp
@@ -10,9 +10,9 @@ using namespace tiny;
 uint16_t Crc16::byte(uint16_t seed, uint8_t byte)
 {
   auto crc = seed;
-  byte = crc >> 8 ^ byte;
-  byte ^= byte >> 4;
-  return (crc << 8) ^ (static_cast<uint16_t>(byte << 12)) ^ (static_cast<uint16_t>(byte << 5)) ^ (static_cast<uint16_t>(byte));
+  byte = static_cast<uint8_t>(crc >> 8 ^ byte);
+  byte ^= static_cast<uint8_t>(byte >> 4);
+  return static_cast<uint16_t>((crc << 8) ^ (static_cast<uint16_t>(byte << 12)) ^ (static_cast<uint16_t>(byte << 5)) ^ (static_cast<uint16_t>(byte)));
 }
 
 uint16_t Crc16::block(uint16_t seed, const uint8_t* bytes, size_t byte_count)
